Add Command to link notes, and code consistency fixes
Changes made:
Add a new feature to link notes. link notes is done by link
<note_ID_1> <note_ID_2>. Will throw various exceptions if not in the
correct format. Also, user is not allowed to link a note to itself
(circular referencing).
Add a hashset for every note that stores all the notes that it is
linkedTo, and all the notes that it is linkedBy.
Minor fixes to coding quality standard, proper exception handling
for existing commands, for TagCommand and DeleteNoteCommand.
Minor fixes to coding quality standard for parser such that all methods
in parser will take in the string array format inputs instead of input
to ensure consistency.
All the relevant JUnit tests have been also updated to pass.