Add ListLinkedNotedCommand
ListLinkedNotesCommand can either, based on the user input, list all the
notes that is pointing to the current note, or all the notes that the
current note is pointing to.

Note: Did not sort the notes when listing. Not sure if it is needed or
useful to sort the notes based on creation date.